Designed the interface in reference to the sketch.
Mostly because we're meeting with Stuart tomorrow, I've also coded the
frames for accessing the assets and previewing the video. Because they
were styled to be fixed, they won't shift when you scroll through the
timeline.

I'm not sure if I'm happy with the interface, but I'm sure I'll continue
to redesign it as we work through the project. Until then, we'll just
ignore their designations.