Began scripting the functionality of the timeline.
At the moment, the script instantiates an asset into the timeline in it's
second timetrack. The information of the asset is provided from JSON,
which will eventually be returned from AJAX.

The asset is styled with a repeating background from the video, which it
assumes is formatted as a PNG. It is also styled with a rounded border,
which required the timetracks to be resized a bit.