Attempted to add the code for jumping your assets.
Ugh! This is still really buggy! The asset does not jump any more than
once to avoid illegal positions, the cursor is not updated in the
statusbar, and the algorith for managing the direction is absolutely
horrible.

I'm going to get myself some dinner; maybe I'll fix it up after I've eaten
a bit.