Finally finished the functionality for jumping.
Phew! I'm still not confident in my implementations of the prototyping
paradigm supported in javascript, but after restructuring all my objects,
I am much more satisfied with my coding.

After jumping, the asset will be shifted in both position and order. The
statusbar will remain updated by increasing or decreasing the cursor. If
the asset attempts to jump between other assets where it does not fit, it
will jump again.

I removed the hack for handling the illegal positions; the process is
currently too complicated to abrogate. If the asset attempts to jump off
the left of the screen, it will glitch out. :<